backend/requirements.txt
pyrofork
tgcrypto
aiohttp
motor
dnspython
jinja2
python-dotenv
